requirements.txt
docx==0.2.4
openai==1.36.1
pandas==2.2.2
pdfminer==20191125
python-docx==1.1.2
python-dotenv==1.0.1
